\name{cell.gene}
\alias{cell.gene.rda}
\docType{data}
\title{cell-gene key table}
\description{
  This is the cell-gene key table: mapping between liver, brain, and lung to their respective tissue specific markers genes.  
  This was obtained from Tissue Specific Gene Expression Regulation (TIGER) database and GENENOTE
}
\format{
	A data matrix of two columns and 237 rows of marker genes.
}
\usage{
data('cell.gene')
}
\references{
    Liu X, Yu X, Zack DJ, Zhu H, Qian J: TiGER: a database for tissue-specific gene expression and regulation. BMC Bioinformatics 2008, 9:271 
	
	Yanai I, Benjamin H, Shmoish M, Chalifa-Caspi V, Shklar M, Ophir R, Bar-Even A, Horn-Saban S, Safran M, Domany E, Lancet D, Shmueli O: Genome-wide midrange transcription profiles reveal expression level relationships in human tissue specification. Bioinformatics 2005, 21:650–659.
}
