% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_names.R
\name{del_names}
\alias{del_names}
\title{Delete name of xts}
\usage{
del_names(x)
}
\arguments{
\item{x}{xts time series}
}
\description{
Delete name of xts
}
\details{
This function can be helpful if one xts is created to be equal to another xts and then changed afterwards. In these cases the new xts inherits the column name of the old xts.
}
\examples{
timeseries <- dsa::daily_sim()$original # timeseries inherits name from original
colnames(timeseries)
colnames(del_names(timeseries))
y <- del_names(timeseries)
colnames(merge(timeseries, y))
}
\author{
Daniel Ollech
}
