% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2HTML.R
\name{df2HTML}
\alias{df2HTML}
\title{Output a dataframe to HTML}
\usage{
df2HTML(dataframe, file)
}
\arguments{
\item{dataframe}{data to be parsed to HTML}

\item{file}{path to save to}
}
\description{
Output a dataframe to a HTML file.
}
\details{
Function used to create HTML for the results of the seasonal adjustment. But can basically be used to create HTML output for any data.frame.
}
\examples{
a=data.frame(lapply(1:4, function(x) round(rnorm(10),2)))
colnames(a) = paste0("x", 1:4)
df2HTML(a, "out.html")
}
\author{
Daniel Ollech
}
