% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayOfMonthDummy.R
\name{dom_dummy}
\alias{dom_dummy}
\title{Dummy for the Day of the Month}
\usage{
dom_dummy(day = "01", start = "2010/1/1", end = "2015/01/01",
  delete29 = T)
}
\arguments{
\item{day}{Day of the Month for which dummy is created}

\item{start}{Startdate}

\item{end}{Enddate}

\item{delete29}{Should the 29th of February be deleted?}
}
\description{
Creates dummies for each chosen day of the week.
}
\details{
This function is used in dsa() to create day of the month dummies.
}
\examples{
plot(dom_dummy())
}
\author{
Daniel Ollech
}
