% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_dummy.R
\name{doy_dummy}
\alias{doy_dummy}
\title{Dummy for the Day of the Year}
\usage{
doy_dummy(day = "1", start = "2010/1/1", end = "2015/01/01")
}
\arguments{
\item{day}{Day of the year for which dummy is created}

\item{start}{Startdate}

\item{end}{Enddate}
}
\description{
Creates dummies for each chosen day of the year
}
\details{
This function is used in dsa() to create day of the year dummies.
}
\examples{
plot(doy_dummy())
}
\author{
Daniel Ollech
}
