% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sa.R
\name{get_sa}
\alias{get_sa}
\title{Get Seasonally Adjusted Series}
\usage{
get_sa(daily.object, forecast = FALSE)
}
\arguments{
\item{daily.object}{Output from dsa}

\item{forecast}{Include forecast of component}
}
\description{
Get the calendar- and seasonally adjusted series from a seasonal adjustment object created by the dsa function
}
\examples{
set.seed(123)
x = daily_sim(n=4)$original # series with length 4 years
res <- dsa(x, cval=7, model=c(3,1,0),fourier_number = 13)
get_sa(res)
}
\seealso{
get_trend, get_original
}
\author{
Daniel Ollech
}
