% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holiday.R
\name{Holiday}
\alias{Holiday}
\title{Creating Holiday dummy}
\usage{
Holiday(dates = timeDate::Easter(2000:2030), shift = 0)
}
\arguments{
\item{dates}{Holiday and period for which dummy shall be created}

\item{shift}{shifting point in time for dummy}
}
\description{
This function uses the Holiday dates of the timeDate::timeDate package to create dummies on a specified holiday.
}
\details{
With shift the user can specify for how many days before (negative value) or after (positive value) the holiday a dummy will be created.
}
\examples{
Holiday(dates=timeDate::Easter(2000:2030), shift=-1)
}
\author{
Daniel Ollech
}
