% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cal.R
\name{make_cal}
\alias{make_cal}
\title{Creating holiday regressor that increases linearly up to holiday and decreases afterwards}
\usage{
make_cal(holidays = NULL, h = 365, original = NA, original2 = NA)
}
\arguments{
\item{holidays}{Holidays for which regressor will be created}

\item{h}{Forecast horizon}

\item{original}{xts time series which characteristics will be used}

\item{original2}{ts time series which characteristics will be used}
}
\description{
Creating holiday regressor that increases linearly up to holiday and decreases afterwards
}
\details{
This function is used internally in dsa()
}
\examples{
a <- daily_sim(n=8)$original
\dontrun{make_cal(holidays="Easter", original=a, original2=xts2ts(a, freq=365))}
}
\author{
Daniel Ollech
}
