% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummy.R
\name{makeDummy}
\alias{makeDummy}
\title{Creating set of dummy variables for specified Holidays}
\usage{
makeDummy(
  holidays = NULL,
  from = -5,
  to = 5,
  h = 365,
  original = NA,
  original2 = NA
)
}
\arguments{
\item{holidays}{holidays for which dummy variables will be created}

\item{from}{start of holiday regressor. Relative to specified holiday}

\item{to}{end of holiday regressor. Relative to specified holiday}

\item{h}{forecast horizon}

\item{original}{xts time series which characteristics will be used}

\item{original2}{ts time series which characteristics will be used}
}
\description{
Creating set of dummy variables for specified Holidays
}
\details{
This function is used internally in dsa()
}
\author{
Daniel Ollech
}
