% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_xts2ts.R
\name{multi_xts2ts}
\alias{multi_xts2ts}
\title{Change multiple xts to a multivariate ts}
\usage{
multi_xts2ts(x, short = FALSE)
}
\arguments{
\item{x}{xts time series}

\item{short}{Is series too short for xts2ts to work?}
}
\description{
Change multiple xts to a multivariate ts
}
\details{
If the ts are used for forecasting
}
\examples{
x <- dsa::daily_sim()$original
y <- dsa::daily_sim()$original
multi_xts2ts(merge(x,y))
}
\author{
Daniel Ollech
}
