% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericFunctions.R
\name{print.daily}
\alias{print.daily}
\title{Print daily time series}
\usage{
\method{print}{daily}(x, ...)
}
\arguments{
\item{x}{Result of dsa() that will be printed}

\item{...}{further arguments handed to print()}
}
\description{
Print output for objects of class "daily"
}
\examples{
x <- daily_sim(3)$original
\dontrun{res<- dsa(x, fourier_number = 24, outlier.types="AO", reg.create=NULL, model=c(3,1,0))}
\dontrun{print(res)}
}
\author{
Daniel Ollech
}
