\name{print.DSA}
\alias{print.DSA}
\title{prints a DSA object}
\description{
  Prettily prints a DSA object.
}
\usage{
print.DSA(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object returned from a call to the DSA routine. }
  \item{\dots}{ Currently ignored }
}
\keyword{models}% at least one, from doc/KEYWORDS

