% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaler.R
\name{Scaler}
\alias{Scaler}
\title{Take logs and differences of  a time series}
\usage{
Scaler(x, Diff = 0, Sdiff = 0, Log = FALSE)
}
\arguments{
\item{x}{time series}

\item{Diff}{number of differences to be taken}

\item{Sdiff}{number of seasonal differences to be taken}

\item{Log}{Should time series be logarithmised}
}
\description{
Logarithmise and / or difference a time series
}
\details{
Function is used in dsa to let the user decide whether logs and differences should be taken.
}
\examples{
a = ts(rnorm(100, 100, 10), start=c(2015,1), frequency=12)
Scaler(a, Diff=1, Log=TRUE)
}
\author{
Daniel Ollech
}
