% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time.R
\name{Time}
\alias{Time}
\title{Creating Several Holiday dummy}
\usage{
Time(from = -10, to = 10, dates = timeDate::Easter(2000:2030))
}
\arguments{
\item{from}{Relative to Holiday, starting date}

\item{to}{Relative to Holiday, end date}

\item{dates}{Which Holidays shall be used}
}
\description{
This function uses the Holiday dates of the timeDate package to create several dummies on a specified holiday.
}
\details{
With shift the user can specify for how many days before (negative value) or after (positive value) the holiday a dummy will be created.
}
\examples{
\dontrun{output(Time(from=5, to=10, dates=timeDate::Easter(2000:2030))}
}
\author{
Daniel Ollech
}
