% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2xts.R
\name{ts2xts}
\alias{ts2xts}
\title{Change ts to xts}
\usage{
ts2xts(x_ts)
}
\arguments{
\item{x_ts}{ts series to be changed to xts}
}
\description{
Change the format of a time series from ts to xts. Has been optimised for the use in dsa(), i.e. for daily time series.
}
\details{
This function is used internally in dsa(). Does not create values for the 29th of February.
}
\examples{
ts2xts(stats::ts(rnorm(1000, 10,1), start=c(2001,1), freq=365))
}
\author{
Daniel Ollech
}
