% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_sum.R
\name{ts_sum}
\alias{ts_sum}
\title{Add time series}
\usage{
ts_sum(...)
}
\arguments{
\item{...}{list of ts time series that are added together}
}
\description{
Sequentially add a set of time series
}
\details{
This function is used internally in dsa()
}
\examples{
ts_sum(list(ts(rnorm(100,10,1)), ts(rnorm(100,10,1)), ts(rnorm(100,10,1))))
}
\author{
Daniel Ollech
}
