\name{weight}
\alias{weight.rda}
\docType{data}
\title{weight data matrix}
\description{
  This is a sample matrix for the weight proportions for three cell types: lung, liver, brain.  
  The proportions of each specific cell type is in columns and samples in rows.
}
\format{
	A data matrix with 3 columns and 14 rows.
}
\usage{
data('weight')
}
