% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts2ts.R
\name{xts2ts}
\alias{xts2ts}
\title{Change xts to ts}
\usage{
xts2ts(series, freq = NULL)
}
\arguments{
\item{series}{xts series to be changed to ts}

\item{freq}{frequency of ts series}
}
\description{
Change the format of a time series from xts to ts. Has been optimised for the use in dsa(), i.e. for daily time series.
}
\details{
This function is used internally in dsa(). Does not create values for the 29th of February.
}
\examples{
xts2ts(xts::xts(rnorm(1095, 10,1), seq.Date(as.Date("2010-01-01"), length.out=1095, by="days")))
}
\author{
Daniel Ollech
}
