% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot the periodogram of a daily time series}
\usage{
plot_spectrum(
  x,
  xlog = FALSE,
  size = 1,
  color = "black",
  vline_color = "#6F87B2"
)
}
\arguments{
\item{x}{xts or ts, daily timeseries}

\item{xlog}{should x-axis be log transformed}

\item{size}{linesize}

\item{color}{color of line}

\item{vline_color}{color of vertical lines}
}
\description{
Plot the periodogram of a daily time series
}
\details{
Plot uses ggplot2 and can be changed accordingly. The spectrum is build around the spec.pgram() function
}
\examples{
x <- daily_sim(3)$original
plot_spectrum(x)
}
\author{
Daniel Ollech
}
